requirements.txt
tensorflow
pytorch
sklearn
tqdm
matplotlib
python-telegram-bot